增加了修改学分兑换率
li

+++ src/main/java/com/attendance/utils/SecurityUtils.java
@@ -19,7 +19,7 @@ public class SecurityUtils
     /**
      * 用户ID
      **/
-    public static Long getUserId()
+    public static Integer getUserId()
     {
         try
         {
